chore: extract common video export functionality to a mixin
