+++ src/pupil_labs/neon_player/plugins/shared/__init__.py
@@ -1,0 +1,3 @@
+from pupil_labs.neon_player.plugins.shared.video_export import BackgroundVideoExportMixin
+
+__all__ = ["BackgroundVideoExportMixin"]
